tie alias creation to premium subscriptions
